Saves an Excel (xlsx) file with metadata templates

This function saves a XLSX file with metadata template to the specified location.

Usage

save_metadata_templates(path = "metadata_template.xlsx")

Arguments

path

File path where the XLSX file with templates will be saved. If left empty (default), the file will be saved in the current working directory under the file "metadata_template.xlsx"